+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * golang 1.18 is now the minimum version
 * add function `OneOfSliceWith` with generic parameters in package `basiccheck` to replace `OneOfStringsWith` function which is now deprecated
+* add function `InSlice` with generic parameters in package `basiccheck` to replace `StringInSlice`, `IntInSlice`, `Int64InSlice` functions which are now deprecated
 
 ## v0.4.1